+++ src/walk.ts
@@ -1,5 +1,11 @@
 import { asyncWalk, walk } from 'estree-walker'
-import { isExpressionType, isFunctionType } from './check'
+import {
+  isExpressionType,
+  isForStatement,
+  isFunctionType,
+  isIdentifier,
+  isReferenced,
+} from './check'
 import { extractIdentifiers } from './extract'
 import { resolveString } from './resolve'
 import { TS_NODE_TYPES } from './utils'
@@ -315,10 +321,6 @@ export function walkIdentifiers(
   })
 }
 
-function isIdentifier(node: t.Node): node is t.Identifier {
-  return node.type === 'Identifier' || node.type === 'JSXIdentifier'
-}
-
 export function isReferencedIdentifier(
   id: t.Identifier,
   parent: t.Node | null | undefined,
@@ -420,16 +422,6 @@ export function walkBlockDeclarations(
   }
 }
 
-function isForStatement(
-  stmt: t.Node,
-): stmt is t.ForStatement | t.ForOfStatement | t.ForInStatement {
-  return (
-    stmt.type === 'ForOfStatement' ||
-    stmt.type === 'ForInStatement' ||
-    stmt.type === 'ForStatement'
-  )
-}
-
 function walkForStatement(
   stmt: t.ForStatement | t.ForOfStatement | t.ForInStatement,
   isVar: boolean,
@@ -468,187 +460,4 @@ function markScopeIdentifier(
   }
   markKnownIds(name, knownIds)
   ;(node.scopeIds || (node.scopeIds = new Set())).add(name)
-}
-
-/**
- * Copied from https://github.com/babel/babel/blob/main/packages/babel-types/src/validators/isReferenced.ts
- * To avoid runtime dependency on @babel/types (which includes process references)
- * This file should not change very often in babel but we may need to keep it
- * up-to-date from time to time.
- *
- * https://github.com/babel/babel/blob/main/LICENSE
- *
- */
-function isReferenced(
-  node: t.Node,
-  parent: t.Node,
-  grandparent?: t.Node,
-): boolean {
-  switch (parent.type) {
-    // yes: PARENT[NODE]
-    // yes: NODE.child
-    // no: parent.NODE
-    case 'MemberExpression':
-    case 'OptionalMemberExpression':
-      if (parent.property === node) {
-        return !!parent.computed
-      }
-      return parent.object === node
-
-    case 'JSXMemberExpression':
-      return parent.object === node
-    // no: let NODE = init;
-    // yes: let id = NODE;
-    case 'VariableDeclarator':
-      return parent.init === node
-
-    // yes: () => NODE
-    // no: (NODE) => {}
-    case 'ArrowFunctionExpression':
-      return parent.body === node
-
-    // no: class { #NODE; }
-    // no: class { get #NODE() {} }
-    // no: class { #NODE() {} }
-    // no: class { fn() { return this.#NODE; } }
-    case 'PrivateName':
-      return false
-
-    // no: class { NODE() {} }
-    // yes: class { [NODE]() {} }
-    // no: class { foo(NODE) {} }
-    case 'ClassMethod':
-    case 'ClassPrivateMethod':
-    case 'ObjectMethod':
-      if (parent.key === node) {
-        return !!parent.computed
-      }
-      return false
-
-    // yes: { [NODE]: "" }
-    // no: { NODE: "" }
-    // depends: { NODE }
-    // depends: { key: NODE }
-    case 'ObjectProperty':
-      if (parent.key === node) {
-        return !!parent.computed
-      }
-      // parent.value === node
-      return !grandparent || grandparent.type !== 'ObjectPattern'
-    // no: class { NODE = value; }
-    // yes: class { [NODE] = value; }
-    // yes: class { key = NODE; }
-    case 'ClassProperty':
-    case 'ClassAccessorProperty':
-      if (parent.key === node) {
-        return !!parent.computed
-      }
-      return true
-    case 'ClassPrivateProperty':
-      return parent.key !== node
-
-    // no: class NODE {}
-    // yes: class Foo extends NODE {}
-    case 'ClassDeclaration':
-    case 'ClassExpression':
-      return parent.superClass === node
-
-    // yes: left = NODE;
-    // no: NODE = right;
-    case 'AssignmentExpression':
-      return parent.right === node
-
-    // no: [NODE = foo] = [];
-    // yes: [foo = NODE] = [];
-    case 'AssignmentPattern':
-      return parent.right === node
-
-    // no: NODE: for (;;) {}
-    case 'LabeledStatement':
-      return false
-
-    // no: try {} catch (NODE) {}
-    case 'CatchClause':
-      return false
-
-    // no: function foo(...NODE) {}
-    case 'RestElement':
-      return false
-
-    case 'BreakStatement':
-    case 'ContinueStatement':
-      return false
-
-    // no: function NODE() {}
-    // no: function foo(NODE) {}
-    case 'FunctionDeclaration':
-    case 'FunctionExpression':
-      return false
-
-    // no: export NODE from "foo";
-    // no: export * as NODE from "foo";
-    case 'ExportNamespaceSpecifier':
-    case 'ExportDefaultSpecifier':
-      return false
-
-    // no: export { foo as NODE };
-    // yes: export { NODE as foo };
-    // no: export { NODE as foo } from "foo";
-    case 'ExportSpecifier':
-      // @ts-expect-error todo(flow->ts): Property 'source' does not exist on type 'AnyTypeAnnotation'.
-      if (grandparent?.source) {
-        return false
-      }
-      return parent.local === node
-
-    // no: import NODE from "foo";
-    // no: import * as NODE from "foo";
-    // no: import { NODE as foo } from "foo";
-    // no: import { foo as NODE } from "foo";
-    // no: import NODE from "bar";
-    case 'ImportDefaultSpecifier':
-    case 'ImportNamespaceSpecifier':
-    case 'ImportSpecifier':
-      return false
-
-    // no: import "foo" assert { NODE: "json" }
-    case 'ImportAttribute':
-      return false
-
-    // no: <div NODE="foo" />
-    case 'JSXAttribute':
-      return false
-
-    // no: [NODE] = [];
-    // no: ({ NODE }) = [];
-    case 'ObjectPattern':
-    case 'ArrayPattern':
-      return false
-
-    // no: new.NODE
-    // no: NODE.target
-    case 'MetaProperty':
-      return false
-
-    // yes: type X = { someProperty: NODE }
-    // no: type X = { NODE: OtherType }
-    case 'ObjectTypeProperty':
-      return parent.key !== node
-
-    // yes: enum X { Foo = NODE }
-    // no: enum X { NODE }
-    case 'TSEnumMember':
-      return parent.id !== node
-
-    // yes: { [NODE]: value }
-    // no: { NODE: value }
-    case 'TSPropertySignature':
-      if (parent.key === node) {
-        return !!parent.computed
-      }
-
-      return true
-  }
-
-  return true
 }
